feat(deps): Complete package migration for CAAPP-430
This commit finalizes the migration of the last set of packages referenced in ticket CAAPP-430. The updated packages include point_castle, provider, shared_preferences, and uni_links2.

The migration of point_castle is currently blocked at version 3.7.0. This is due to a dependency constraint from the 'encrypt' package, which is no longer maintained and does not support newer versions of point_castle. As a future improvement, it is recommended to replace 'encrypt' with the 'cryptography' package to allow for further updates.

The uni_links2 package has been updated to its latest version. However, this package is also unmaintained. The official recommendation is to migrate to 'applinks'. An attempt to switch to 'applinks' was made, but it introduced a chain of dependencies that would require a Gradle update. This migration should be revisited after the Gradle update is complete.

Closes: CAAPP-430